algebra
Module algebra

Traits describing algebras, and implementations thereof for different representations of the Steenrod algebra.

Re-exports

pub use adem_algebra::AdemAlgebra;

pub use adem_algebra::AdemAlgebraT;

pub use algebra_trait::Algebra;

pub use algebra_trait::GeneratedAlgebra;

pub use algebra_trait::MuAlgebra;

pub use algebra_trait::UnstableAlgebra;

pub use bialgebra_trait::Bialgebra;

pub use field::Field;

pub use milnor_algebra::MilnorAlgebra;

pub use milnor_algebra::MilnorAlgebraT;

pub use polynomial_algebra::PolynomialAlgebra;

pub use polynomial_algebra::PolynomialAlgebraMonomial;

pub use polynomial_algebra::PolynomialAlgebraTableEntry;

pub use steenrod_algebra::AlgebraType;

pub use steenrod_algebra::SteenrodAlgebra;

pub use steenrod_algebra::SteenrodAlgebraBorrow;

pub use steenrod_algebra::SteenrodAlgebraT;

Modules

adem_algebra

The Steenrod algebra using the Adem basis.

algebra_trait 🔒

bialgebra_trait 🔒

combinatorics

field

Finite fields over a prime.

milnor_algebra

pair_algebra

This implements the notion of a split pair algebra in the sense of https://arxiv.org/abs/2105.07628v1, whose notation we will use throughout.

polynomial_algebra 🔒

steenrod_algebra 🔒