Function algebra::algebra::adem_algebra::shift_vec

unsafe fn shift_vec<T>(v: &mut Vec<T>, offset: isize)

Shifts a Vec’s elements back by offset.

Safety

This function leaves v in an invalid state, since ptr does not point to the beginning of an allocation after this function completes. Operations which may trigger resizing must be avoided, since the allocator will otherwise get extremely confused.

The sum of offsets to calls to this function for a particular v must always be positive (i.e., to avoid pointing before the allocation), less than the v’s capacity (to avoid pointing after), and zero when v is destroyed or resized (for the reasons described above)