Struct algebra::algebra::adem_algebra::AdemBasisElement

pub struct AdemBasisElement {
    pub degree: i32,
    pub bocksteins: u32,
    pub ps: Vec<u32>,
    pub p_or_sq: bool,
}

An Adem basis element for the Steenrod algebra.

To encode an element $$ \beta^{\varepsilon_0} P^{i_0}
\beta^{\varepsilon_1} P^{i_1} \cdots \beta^{\varepsilon_n} P^{i_n} \beta^{\varepsilon_{n+1}}, $$ we set $$ \begin{aligned} \mathtt{ps} &= [i_0, i_1, \ldots, i_n] \\ \mathtt{bocksteins} &= 000\cdots0\varepsilon_{n+1} \varepsilon_n \cdots \varepsilon_0 \end{aligned} $$

Fields
degree: i32

The degree of the element.

bocksteins: u32

A bitset of which $\beta$ Bocksteins are in the element’s decomposition.

ps: Vec<u32>

A list of which Steenrod powers are in the element’s decompositions.

p_or_sq: bool

Whether to denote the generators as powers $P^i$ or squares $Sq^i$ when using string formatting; true denotes the former.

Implementations

impl AdemBasisElement

fn iter_filtered(&self) -> impl Iterator<Item = PorBockstein> + '_

Returns an iterator over the element’s decomposition.

This returns alternating Bocksteins and Ps, but skipping any \beta^0 factors.

fn excess(&self, p: ValidPrime) -> u32