algebra::algebra::adem_algebra
Trait AdemAlgebraT

pub trait AdemAlgebraT:
    Send
    + Sync
    + Algebra {
    // Required method
    fn adem_algebra(&self) -> &AdemAlgebra;
}

An algebra that can be viewed as an Adem algebra.

This is here so that the Python bindings can use modules defined for AdemAlgebraT with their version of SteenrodAlgebra.

In order for things to work AdemAlgebraT cannot implement Algebra. Otherwise, the algebra enum for our bindings will see an implementation clash.

Required Methods

fn adem_algebra(&self) -> &AdemAlgebra