Trait algebra::algebra::algebra_trait::Algebra

pub trait Algebra: Display + Send + Sync + 'static {
    fn prime(&self) -> ValidPrime;

    fn compute_basis(&self, degree: i32);

    fn dimension(&self, degree: i32, excess: i32) -> usize;

    fn multiply_basis_elements(
        &self, 
        result: SliceMut<'_>, 
        coeff: u32, 
        r_degree: i32, 
        r_idx: usize, 
        s_degree: i32, 
        s_idx: usize, 
        excess: i32
    );

    fn basis_element_to_string(&self, degree: i32, idx: usize) -> String;

    fn multiply_basis_element_by_element(
        &self, 
        result: SliceMut<'_>, 
        coeff: u32, 
        r_degree: i32, 
        r_idx: usize, 
        s_degree: i32, 
        s: Slice<'_>, 
        excess: i32
    ) { ... }

    fn multiply_element_by_basis_element(
        &self, 
        result: SliceMut<'_>, 
        coeff: u32, 
        r_degree: i32, 
        r: Slice<'_>, 
        s_degree: i32, 
        s_idx: usize, 
        excess: i32
    ) { ... }

    fn multiply_element_by_element(
        &self, 
        result: SliceMut<'_>, 
        coeff: u32, 
        r_degree: i32, 
        r: Slice<'_>, 
        s_degree: i32, 
        s: Slice<'_>, 
        excess: i32
    ) { ... }

    fn default_filtration_one_products(&self) -> Vec<(String, i32, usize)> { ... }

    fn element_to_string(&self, degree: i32, element: Slice<'_>) -> String { ... }
}

A graded algebra over F_p, finite dimensional in each degree, equipped with a choice of ordered basis in each dimension. Basis elements of the algebra are referred to by their degree and index, and general elements are referred to by the degree and an FpVector listing the coefficients of the element in terms of the basis.

Since the graded algebra is often infinite dimensional, we cannot construct a complete description of the algebra. Instead, we use the function compute_basis(degree). When called, the algebra should compute relevant data to be able to perform calculations up to degree degree. It is the responsibility of users to ensure compute_degree(degree) is called before calling other functions with the degree parameter.

The algebra should also come with a specified choice of algebra generators, which are necessarily basis elements. It gives us a simpler way of describing finite modules by only specifying the action of the generators.

Required methods

fn prime(&self) -> ValidPrime

Returns the prime the algebra is over.

fn compute_basis(&self, degree: i32)

Computes the list of basis elements up to and including degree degree. This should include any other preparation needed to evaluate all the other functions that involve a degree parameter. One should be able to call compute_basis multiple times, and there should be little overhead when calling compute_basis(degree) multiple times with the same degree.

fn dimension(&self, degree: i32, excess: i32) -> usize

Gets the dimension of the algebra in degree degree.

fn multiply_basis_elements(
&self,
result: SliceMut<'_>,
coeff: u32,
r_degree: i32,
r_idx: usize,
s_degree: i32,
s_idx: usize,
excess: i32
)

Computes the product r * s of the two basis elements, and adds the result to result.

result is not required to be aligned.

fn basis_element_to_string(&self, degree: i32, idx: usize) -> String

Converts a basis element into a string for display.

Loading content...
Provided methods

fn multiply_basis_element_by_element(
&self,
result: SliceMut<'_>,
coeff: u32,
r_degree: i32,
r_idx: usize,
s_degree: i32,
s: Slice<'_>,
excess: i32
)

result and s are not required to be aligned.

fn multiply_element_by_basis_element(
&self,
result: SliceMut<'_>,
coeff: u32,
r_degree: i32,
r: Slice<'_>,
s_degree: i32,
s_idx: usize,
excess: i32
)

result and r are not required to be aligned.

fn multiply_element_by_element(
&self,
result: SliceMut<'_>,
coeff: u32,
r_degree: i32,
r: Slice<'_>,
s_degree: i32,
s: Slice<'_>,
excess: i32
)

result, r and s are not required to be aligned.

fn default_filtration_one_products(&self) -> Vec<(String, i32, usize)>

A filtration one element in Ext(k, k) is the same as an indecomposable element of the algebra. This function returns a default list of such elements in the format (name, degree, index) for whom we want to compute products with in the resolutions.

fn element_to_string(&self, degree: i32, element: Slice<'_>) -> String

Converts an element into a string for display.

Loading content...