Trait algebra::algebra::algebra_trait::GeneratedAlgebra

pub trait GeneratedAlgebra: Algebra {
    fn generators(&self, degree: i32) -> Vec<usize>;

    fn string_to_generator<'a, 'b>(
        &'a self, 
        input: &'b str
    ) -> IResult<&'b str, (i32, usize)>;

    fn decompose_basis_element(
        &self, 
        degree: i32, 
        idx: usize
    ) -> Vec<(u32, (i32, usize), (i32, usize))>;

    fn generating_relations(
        &self, 
        degree: i32
    ) -> Vec<Vec<(u32, (i32, usize), (i32, usize))>>;

    fn generator_to_string(&self, degree: i32, idx: usize) -> String { ... }
}

An algebra with a specified list of generators and generating relations. This data can be used to specify modules by specifying the actions of the generators.

Required methods

fn generators(&self, degree: i32) -> Vec<usize>

Given a degree degree, the function returns a list of algebra generators in that degree. This return value is the list of indices of the basis elements that are generators. The list need not be in any particular order.

This method need not be fast, because they will only be performed when constructing the module, and will often only involve low dimensional elements.

fn string_to_generator<'a, 'b>(
&'a self,
input: &'b str
) -> IResult<&'b str, (i32, usize)>

This parses a string and returns the generator described by the string. The signature of this function is the same nom combinators.

This function MUST be inverse to string_to_generator (and not basis_element_to_string).

fn decompose_basis_element(
&self,
degree: i32,
idx: usize
) -> Vec<(u32, (i32, usize), (i32, usize))>

Given a non-generator basis element of the algebra, decompose it in terms of algebra generators. Recall each basis element is given by a pair $(d, i))$, where $d$ is the degree of the generator, and $i$ is the index of the basis element. Given a basis element $A$, the function returns a list of triples $(c_i, A_i, B_i)$ where each $A_i$ and $B_i$ are basis elements of strictly smaller degree than the original, and $$ A = \sum_i c_i A_i B_i.$$ This allows us to recursively compute the action of the algebra.

This method need not be fast, because they will only be performed when constructing the module, and will often only involve low dimensional elements.

fn generating_relations(
&self,
degree: i32
) -> Vec<Vec<(u32, (i32, usize), (i32, usize))>>

Get any relations that the algebra wants checked to ensure the consistency of module.

Loading content...
Provided methods

fn generator_to_string(&self, degree: i32, idx: usize) -> String

This returns the name of a generator. Note that the index is the index of the generator in the list of all basis elements. It is undefined behaviour to call this function with a (degree, index) pair that is not a generator.

The default implementation calls self.basis_element_to_string, but occassionally the generators might have alternative, more concise names that are preferred.

This function MUST be inverse to string_to_generator.

Loading content...