Trait algebra::algebra::algebra_trait::JsonAlgebra

pub trait JsonAlgebra: Algebra {
    fn prefix(&self) -> &str;

    fn json_to_basis(&self, _json: Value) -> Result<(i32, usize)>;

    fn json_from_basis(&self, _degree: i32, _idx: usize) -> Value;
}

Required methods

fn prefix(&self) -> &str

fn json_to_basis(&self, _json: Value) -> Result<(i32, usize)>

Converts a JSON object into a basis element. The way basis elements are represented by JSON objects is to be specified by the algebra itself, and will be used by module specifications.

fn json_from_basis(&self, _degree: i32, _idx: usize) -> Value

Loading content...