Module algebra::algebra::milnor_algebra

Structs

Matrix2D

MilnorAlgebra

MilnorBasisElement

MilnorProfile

PPartAllocation
The parts of a PPartMultiplier that involve heap allocation. This lets us reuse the allocation across multiple different multipliers. Reusing the whole PPartMultiplier is finicky but doable due to lifetime issues, but it appears to be less performant.

PPartMultiplier

QPart

Constants

ZERO_QPART

Traits

MilnorAlgebraT

Type Definitions

PPart

PPartEntry