algebra::algebra::motivic::milnor

Trait Bigraded

pub trait Bigraded: Copy + Debug {
    // Required method
    fn bidegree(self) -> (i32, i32);
}

A monomial carrying a motivic bidegree, in whichever of the two dual conventions its type stands for.

Required Methods

fn bidegree(self) -> (i32, i32)

The (topological degree, motivic weight).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".