Trait algebra::algebra::pair_algebra::PairAlgebra

pub trait PairAlgebra: Algebra {
    type Element;
    fn new_pair_element(&self, degree: i32) -> Self::Element;

    fn sigma_multiply_basis(
        &self, 
        result: &mut Self::Element, 
        coeff: u32, 
        r_degree: i32, 
        r_idx: usize, 
        s_degree: i32, 
        s_idx: usize
    );

    fn a_multiply(
        &self, 
        result: SliceMut<'_>, 
        coeff: u32, 
        r_degree: i32, 
        r: Slice<'_>, 
        s_degree: i32, 
        s: &Self::Element
    );

    fn sigma_multiply(
        &self, 
        result: &mut Self::Element, 
        coeff: u32, 
        r_degree: i32, 
        r: Slice<'_>, 
        s_degree: i32, 
        s: Slice<'_>
    ) { ... }
}

A lift of an algebra to a split pair algebra. See module introduction for more.

Associated Types

type Element

An element in the cohomological degree zero part of the pair algebra. This tends to not be a ring over Fp, so we let the algebra specify how it wants to represent the elements.

Required methods

fn new_pair_element(&self, degree: i32) -> Self::Element

Create a new zero element in the given degree.

fn sigma_multiply_basis(
&self,
result: &mut Self::Element,
coeff: u32,
r_degree: i32,
r_idx: usize,
s_degree: i32,
s_idx: usize
)

Given $r, s \in \pi_0(A)$, compute $\sigma(r) \sigma(s)$ and add the result to result.

fn a_multiply(
&self,
result: SliceMut<'_>,
coeff: u32,
r_degree: i32,
r: Slice<'_>,
s_degree: i32,
s: &Self::Element
)

Compute $A(r, s)$ and write the result to result.

Provided methods

fn sigma_multiply(
&self,
result: &mut Self::Element,
coeff: u32,
r_degree: i32,
r: Slice<'_>,
s_degree: i32,
s: Slice<'_>
)

Same as PairAlgebra::sigma_multiply_basis but with non-basis elements.