Module algebra::alloc::prelude::v1

🔬 This is a nightly-only experimental API. (alloc_prelude)

The first version of the prelude of alloc crate.

See the module-level documentation for more.

Structs

Box
A pointer type for heap allocation.

String
A UTF-8 encoded, growable string.

Vec
A contiguous growable array type, written Vec<T> but pronounced 'vector'.

Traits

ToOwned
A generalization of Clone to borrowed data.

ToString
A trait for converting a value to a String.