Trait algebra::curves::PairingEngine

pub trait PairingEngine: 'static + PartialEq<Self> + Eq + Send + Copy + Sync + Debug where
    <Self::G1Projective as ProjectiveCurve>::BaseField == Self::Fq,
    <Self::G1Projective as ProjectiveCurve>::ScalarField == Self::Fr,
    <Self::G1Projective as ProjectiveCurve>::Affine == Self::G1Affine,
    <Self::G1Affine as AffineCurve>::BaseField == Self::Fq,
    <Self::G1Affine as AffineCurve>::ScalarField == Self::Fr,
    <Self::G1Affine as AffineCurve>::Projective == Self::G1Projective,
    <Self::G2Projective as ProjectiveCurve>::BaseField == Self::Fqe,
    <Self::G2Projective as ProjectiveCurve>::ScalarField == Self::Fr,
    <Self::G2Projective as ProjectiveCurve>::Affine == Self::G2Affine,
    <Self::G2Affine as AffineCurve>::BaseField == Self::Fqe,
    <Self::G2Affine as AffineCurve>::ScalarField == Self::Fr,
    <Self::G2Affine as AffineCurve>::Projective == Self::G2Projective, {
    type Fr: PrimeField + SquareRootField;
    type G1Projective: ProjectiveCurve + Into<Self::G1Affine> + From<Self::G1Affine> + MulAssign<Self::Fr>;
    type G1Affine: AffineCurve + Into<Self::G1Projective> + Into<Self::G1Prepared> + From<Self::G1Projective>;
    type G1Prepared: ToBytes + Clone + From<Self::G1Affine> + Default + Send + Sync + Debug;
    type G2Projective: ProjectiveCurve + Into<Self::G2Affine> + From<Self::G2Affine> + MulAssign<Self::Fr>;
    type G2Affine: AffineCurve + Into<Self::G2Projective> + Into<Self::G2Prepared> + From<Self::G2Projective>;
    type G2Prepared: ToBytes + Clone + From<Self::G2Affine> + Default + Send + Sync + Debug;
    type Fq: PrimeField + SquareRootField;
    type Fqe: SquareRootField;
    type Fqk: Field;
#[must_use]    fn miller_loop<'a, I>(i: I) -> Self::Fqk
    where
        I: IntoIterator<Item = &'a (Self::G1Prepared, Self::G2Prepared)>;

#[must_use]    fn final_exponentiation(&Self::Fqk) -> Option<Self::Fqk>;

#[must_use]    fn product_of_pairings<'a, I>(i: I) -> Self::Fqk
    where
        I: IntoIterator<Item = &'a (Self::G1Prepared, Self::G2Prepared)>,
    { ... }

#[must_use]    fn pairing<G1, G2>(p: G1, q: G2) -> Self::Fqk
    where
        G1: Into<Self::G1Affine>,
        G2: Into<Self::G2Affine>,
    { ... }
}

Associated Types

type Fr: PrimeField + SquareRootField

This is the scalar field of the G1/G2 groups.

type G1Projective: ProjectiveCurve + Into<Self::G1Affine> + From<Self::G1Affine> + MulAssign<Self::Fr>

The projective representation of an element in G1.

type G1Affine: AffineCurve + Into<Self::G1Projective> + Into<Self::G1Prepared> + From<Self::G1Projective>

The affine representation of an element in G1.

type G1Prepared: ToBytes + Clone + From<Self::G1Affine> + Default + Send + Sync + Debug

A G1 element that has been preprocessed for use in a pairing.

type G2Projective: ProjectiveCurve + Into<Self::G2Affine> + From<Self::G2Affine> + MulAssign<Self::Fr>

The projective representation of an element in G2.

type G2Affine: AffineCurve + Into<Self::G2Projective> + Into<Self::G2Prepared> + From<Self::G2Projective>

The affine representation of an element in G2.

type G2Prepared: ToBytes + Clone + From<Self::G2Affine> + Default + Send + Sync + Debug

A G2 element that has been preprocessed for use in a pairing.

type Fq: PrimeField + SquareRootField

The base field that hosts G1.

type Fqe: SquareRootField

The extension field that hosts G2.

type Fqk: Field

The extension field that hosts the target group of the pairing.

Loading content...
Required methods

#[must_use]fn miller_loop<'a, I>(i: I) -> Self::Fqk where
I: IntoIterator<Item = &'a (Self::G1Prepared, Self::G2Prepared)>,

Perform a miller loop with some number of (G1, G2) pairs.

#[must_use]fn final_exponentiation(&Self::Fqk) -> Option<Self::Fqk>

Perform final exponentiation of the result of a miller loop.

Loading content...
Provided methods

#[must_use]fn product_of_pairings<'a, I>(i: I) -> Self::Fqk where
I: IntoIterator<Item = &'a (Self::G1Prepared, Self::G2Prepared)>,

Computes a product of pairings.

#[must_use]fn pairing<G1, G2>(p: G1, q: G2) -> Self::Fqk where
G1: Into<Self::G1Affine>,
G2: Into<Self::G2Affine>,

Performs multiple pairing operations

Loading content...