Module algebra::ed_on_bls12_381

This module implements a twisted Edwards curve whose base field is the scalar field of the curve BLS12-377. This allows defining cryptographic primitives that use elliptic curves over the scalar field of the latter curve. This curve was generated by Sean Bowe, and is also known as Jubjub.

Curve information:

Base field: q = 52435875175126190479447740508185965837690552500527637822603658699938581184513

Scalar field: r = 6554484396890773809930967563523245729705921265872317281365359162392183254199

Valuation(q - 1, 2) = 32

Valuation(r - 1, 2) = 1

Curve equation: ax^2 + y^2 =1 + dx^2y^2, where

a = -1

d = -(10240/10241)

Modules

fq

fr

Structs

EdwardsParameters
JubJub is a twisted Edwards curve. These curves have equations of the form: ax² + y² = 1 - dx²y². over some base finite field Fq.

FqParameters

FrParameters

Type Definitions

EdwardsAffine

EdwardsProjective

Fq

Fr