algebra

Enum MilnorAlgebra

pub enum MilnorAlgebra {
    Polynomial(MilnorAlgebraInner<NoExterior>),
    Exterior(MilnorAlgebraInner<Exterior>),
}

A dual Steenrod algebra in the Milnor basis, of either shape.

Variants

Polynomial(MilnorAlgebraInner<NoExterior>)

Exterior(MilnorAlgebraInner<Exterior>)

Implementations

impl MilnorAlgebra

pub fn has_exterior(&self) -> bool

pub fn q(&self) -> i32

pub fn profile(&self) -> &MilnorProfile

pub fn compute_degree(&self, elt: &mut MilnorBasisElement)

pub fn basis_element_from_index( &self, degree: i32, idx: usize, ) -> MilnorBasisElement

pub fn try_basis_element_to_index( &self, elt: &MilnorBasisElement, ) -> Option<usize>

pub fn basis_element_to_index(&self, elt: &MilnorBasisElement) -> usize

pub fn ppart_table(&self, t: i32) -> &[PPart]

pub fn try_beps_pn(&self, e: u32, x: PPartEntry) -> Option<(i32, usize)>

pub fn beps_pn(&self, e: u32, x: PPartEntry) -> (i32, usize)

pub fn multiply( &self, res: FpSliceMut<'_>, coef: u32, m1: MilnorBasisElement, m2: MilnorBasisElement, )

pub fn multiply_with_allocation( &self, res: FpSliceMut<'_>, coef: u32, m1: MilnorBasisElement, m2: MilnorBasisElement, excess: i32, allocation: PPartAllocation, ) -> PPartAllocation

pub fn new(p: ValidPrime, unstable_enabled: bool) -> Self

The classical dual Steenrod algebra at p.

pub fn new_with_profile( p: ValidPrime, profile: MilnorProfile, unstable_enabled: bool, ) -> Self

The classical dual Steenrod algebra at p, restricted to profile.