Module algebra::fields

Modules

arithmetic

macros

models

utils

Structs

BitIterator

CubicExtField

Fp2ParamsWrapper

Fp3ParamsWrapper

Fp4ParamsWrapper

Fp6ParamsWrapper

Fp12ParamsWrapper

Fp256

Fp320

Fp384

Fp768

Fp832

QuadExtField

Enums

LegendreSymbol

Traits

CubicExtParameters

FftField
The interface for fields that are able to be used in FFTs.

FftParameters
A trait that defines parameters for a field that can be used for FFTs.

Field
The interface for a generic field.

Fp2Parameters

Fp3Parameters

Fp4Parameters

Fp6Parameters

Fp12Parameters

Fp256Parameters

Fp320Parameters

Fp384Parameters

Fp768Parameters

Fp832Parameters

FpParameters
A trait that defines parameters for a prime field.

PrimeField
The interface for a prime field.

QuadExtParameters

SquareRootField
The interface for a field that supports an efficient square-root operation.

Functions

batch_inversion

characteristic_square_mod_6_is_one

Type Definitions

Fp2

Fp3

Fp4

Fp6

Fp12