Trait algebra::fields::models::quadratic_extension::QuadExtParameters

pub trait QuadExtParameters: 'static + Send + Sync {
    type BasePrimeField: PrimeField;
    type BaseField: Field;
    type FrobCoeff: Field;

    const DEGREE_OVER_BASE_PRIME_FIELD: usize;
    const NONRESIDUE: Self::BaseField;
    const FROBENIUS_COEFF_C1: &'static [Self::FrobCoeff];

    fn mul_base_field_by_frob_coeff(fe: &mut Self::BaseField, power: usize);

    fn mul_base_field_by_nonresidue(fe: &Self::BaseField) -> Self::BaseField { ... }

    fn cyclotomic_exp(
        fe: &QuadExtField<Self>, 
        exponent: impl AsRef<[u64]>
    ) -> QuadExtField<Self> { ... }
}

Associated Types

type BasePrimeField: PrimeField

The prime field that this quadratic extension is eventually an extension of.

type BaseField: Field

The base field that this field is a quadratic extension of.

type FrobCoeff: Field

The type of the coefficients for an efficient implemntation of the Frobenius endomorphism.

Loading content...
Associated Constants

const DEGREE_OVER_BASE_PRIME_FIELD: usize

The degree of the extension over the base prime field.

const NONRESIDUE: Self::BaseField

The quadratic non-residue used to construct the extension.

const FROBENIUS_COEFF_C1: &'static [Self::FrobCoeff]

Coefficients for the Frobenius automorphism.

Loading content...
Required methods

fn mul_base_field_by_frob_coeff(fe: &mut Self::BaseField, power: usize)

A specializable method for multiplying an element of the base field by the appropriate Frobenius coefficient.

Loading content...
Provided methods

fn mul_base_field_by_nonresidue(fe: &Self::BaseField) -> Self::BaseField

A specializable method for multiplying an element of the base field by the quadratic non-residue. This is used in Karatsuba multiplication and in complex squaring.

fn cyclotomic_exp(
fe: &QuadExtField<Self>,
exponent: impl AsRef<[u64]>
) -> QuadExtField<Self>

A specializable method for exponentiating that is to be used only when fe is known to be in the cyclotommic subgroup.

Loading content...