Module algebra::module::block_structure

Structs

BlockStructure

A block structure is a structure that makes it efficient to convert between an index into a sum of vector spaces and an index into an individual one. For instance, a FreeModule has some collection of generators in various degrees and a general element is a homogenous sum \sum Op^i x_i. The basis for the FreeModule is divided into a block for each generator x_i. The entries in the block correspond to a basis for the algebra. In order to multiply by a steenrod operation, we have to pull out each block and multiply each block separately by the steenrod operation.

GeneratorBasisEltPair