algebra::module

Module homomorphism

Modules

free_module_homomorphism 🔒

full_module_homomorphism 🔒

generic_zero_homomorphism 🔒

hom_pullback 🔒

quotient_homomorphism 🔒

Structs

FullModuleHomomorphism

A ModuleHomomorphism that simply records the matrix of the homomorphism in every degree. This is currently rather bare bones.

GenericZeroHomomorphism

HomPullback

Given a map $\mathtt{map}: A \to B$ and hom modules $\mathtt{source} = \Hom(B, X)$, $\mathtt{target} = \Hom(A, X)$, produce the induced pullback map $\Hom(B, X) \to \Hom(A, X)$.

MuFreeModuleHomomorphism

QuotientHomomorphism

QuotientHomomorphismSource

Traits

IdentityHomomorphism

ModuleHomomorphism

A trait that represents a homomorphism between two modules.

ZeroHomomorphism

Type Aliases

FreeModuleHomomorphism

UnstableFreeModuleHomomorphism