Struct algebra::module::homomorphism::free_module_homomorphism::FreeModuleHomomorphism

pub struct FreeModuleHomomorphism<M: Module> {
    source: Arc<FreeModule<M::Algebra>>,
    target: Arc<M>,
    outputs: OnceBiVec<Vec<FpVector>>,
    pub images: OnceBiVec<Option<Subspace>>,
    pub kernels: OnceBiVec<Option<Subspace>>,
    pub quasi_inverses: OnceBiVec<Option<QuasiInverse>>,
    min_degree: i32,
    degree_shift: i32,
}

Fields
source: Arc<FreeModule<M::Algebra>>target: Arc<M>outputs: OnceBiVec<Vec<FpVector>>images: OnceBiVec<Option<Subspace>>kernels: OnceBiVec<Option<Subspace>>quasi_inverses: OnceBiVec<Option<QuasiInverse>>min_degree: i32degree_shift: i32

degree shift, such that ouptut_degree = input_degree - degree_shift

Implementations

impl<M: Module> FreeModuleHomomorphism<M>

pub fn new(
source: Arc<FreeModule<M::Algebra>>,
target: Arc<M>,
degree_shift: i32
) -> Self

pub fn degree_shift(&self) -> i32

pub fn min_degree(&self) -> i32

pub fn next_degree(&self) -> i32

pub fn output(&self, generator_degree: i32, generator_index: usize) -> &FpVector

pub fn extend_by_zero(&self, degree: i32)

pub fn add_generators_from_big_vector(
&self,
degree: i32,
outputs_vectors: Slice<'_>
)

pub fn add_generators_from_matrix_rows(
&self,
degree: i32,
matrix: MatrixSliceMut<'_>
)

A MatrixSlice will do but there is no applicaiton of this struct, so it doesn’t exist yet…

pub fn add_generators_from_rows(&self, degree: i32, rows: Vec<FpVector>)

pub fn add_generators_from_rows_ooo(
&self,
degree: i32,
rows: Vec<FpVector>
) -> Range<i32>

Add the image of a bidegree out of order. See OnceVec::push_ooo for details on return value.

pub fn ooo_outputs(&self) -> Vec<i32>

List of outputs that have been added out of order

pub fn apply_to_generator(
&self,
result: &mut FpVector,
coeff: u32,
degree: i32,
idx: usize
)

pub fn set_image(&self, degree: i32, image: Option<Subspace>)

pub fn set_kernel(&self, degree: i32, kernel: Option<Subspace>)

pub fn set_quasi_inverse(
&self,
degree: i32,
quasi_inverse: Option<QuasiInverse>
)

pub fn compose<'a, T: ModuleHomomorphism<Source = M>>(
&'a self,
other: &'a T
) -> CompositionHomomorphism<'_, M, T>

impl<A: Algebra> FreeModuleHomomorphism<FreeModule<A>>

pub fn hom_k(&self, t: i32) -> Vec<Vec<u32>>

Given f: M -> N, compute the dual f*: Hom(N, k) -> Hom(M, k) in source (N) degree t.