Struct algebra::module::homomorphism::BoundedModuleHomomorphism

pub struct BoundedModuleHomomorphism<S: BoundedModule, T: Module<Algebra = S::Algebra>> {
    pub source: Arc<S>,
    pub target: Arc<T>,
    pub degree_shift: i32,
    pub matrices: BiVec<Matrix>,
    pub quasi_inverses: OnceBiVec<QuasiInverse>,
    pub kernels: OnceBiVec<Subspace>,
    pub images: OnceBiVec<Subspace>,
}

Fields
source: Arc<S>target: Arc<T>degree_shift: i32matrices: BiVec<Matrix>quasi_inverses: OnceBiVec<QuasiInverse>kernels: OnceBiVec<Subspace>images: OnceBiVec<Subspace>
Implementations

impl<A, S, T> BoundedModuleHomomorphism<S, T> where
A: Algebra,
S: BoundedModule<Algebra = A>,
T: Module<Algebra = A>,

pub fn new(source: Arc<S>, target: Arc<T>, degree_shift: i32) -> Self

pub fn from_matrices(
source: Arc<S>,
target: Arc<T>,
degree_shift: i32,
matrices: BiVec<Matrix>
) -> Self

pub fn from<F: ModuleHomomorphism<Source = S, Target = T>>(f: &F) -> Self

pub fn replace_source<S_: BoundedModule<Algebra = A>>(
self,
source: Arc<S_>
) -> BoundedModuleHomomorphism<S_, T>

This function replaces the source of the BoundedModuleHomomorphism and does nothing else. This is useful for changing the type of the source (but not the mathematical module itself). This is intended to be used in conjunction with BoundedModule::to_fd_module

pub fn replace_target<T_: BoundedModule<Algebra = A>>(
self,
target: Arc<T_>
) -> BoundedModuleHomomorphism<S, T_>

See replace_source