Type Alias algebra::module::homomorphism::UnstableFreeModuleHomomorphism

pub type UnstableFreeModuleHomomorphism<M> = MuFreeModuleHomomorphism<true, M>;

Aliased Type

struct UnstableFreeModuleHomomorphism<M> {
    source: Arc<MuFreeModule<true, <M as Module>::Algebra>, Global>,
    target: Arc<M, Global>,
    outputs: OnceBiVec<Vec<FpVector, Global>>,
    pub images: OnceBiVec<Option<Subspace>>,
    pub kernels: OnceBiVec<Option<Subspace>>,
    pub quasi_inverses: OnceBiVec<Option<QuasiInverse>>,
    min_degree: i32,
    degree_shift: i32,
}

Fields
source: Arc<MuFreeModule<true, <M as Module>::Algebra>, Global>target: Arc<M, Global>outputs: OnceBiVec<Vec<FpVector, Global>>images: OnceBiVec<Option<Subspace>>kernels: OnceBiVec<Option<Subspace>>quasi_inverses: OnceBiVec<Option<QuasiInverse>>min_degree: i32degree_shift: i32

degree shift, such that ouptut_degree = input_degree - degree_shift

Implementations

impl<const U: bool, M: Module> MuFreeModuleHomomorphism<U, M>where M::Algebra: MuAlgebra<U>,

pub fn new( source: Arc<MuFreeModule<U, M::Algebra>>, target: Arc<M>, degree_shift: i32 ) -> Self

pub fn degree_shift(&self) -> i32

pub fn min_degree(&self) -> i32

pub fn next_degree(&self) -> i32

pub fn output(&self, generator_degree: i32, generator_index: usize) -> &FpVector

pub fn differential_density(&self, degree: i32) -> f32

pub fn extend_by_zero(&self, degree: i32)

pub fn add_generators_from_big_vector( &self, degree: i32, outputs_vectors: Slice<'_> )

pub fn add_generators_from_matrix_rows( &self, degree: i32, matrix: MatrixSliceMut<'_> )

A MatrixSlice will do but there is no applicaiton of this struct, so it doesn’t exist yet…

pub fn add_generators_from_rows(&self, degree: i32, rows: Vec<FpVector>)

pub fn add_generators_from_rows_ooo( &self, degree: i32, rows: Vec<FpVector> ) -> Range<i32>

Add the image of a bidegree out of order. See OnceVec::push_ooo for details on return value.

pub fn ooo_outputs(&self) -> Vec<i32>

List of outputs that have been added out of order

pub fn apply_to_generator( &self, result: &mut FpVector, coeff: u32, degree: i32, idx: usize )

pub fn set_image(&self, degree: i32, image: Option<Subspace>)

pub fn set_kernel(&self, degree: i32, kernel: Option<Subspace>)

pub fn set_quasi_inverse( &self, degree: i32, quasi_inverse: Option<QuasiInverse> )

impl<const U: bool, A: MuAlgebra<U>> MuFreeModuleHomomorphism<U, MuFreeModule<U, A>>

pub fn hom_k(&self, t: i32) -> Vec<Vec<u32>>

Given f: M -> N, compute the dual f*: Hom(N, k) -> Hom(M, k) in source (N) degree t.