Module algebra::prelude

Traits

AffineCurve
Affine representation of an elliptic curve point guaranteed to be in the correct prime order subgroup.

BigInteger
This defines a BigInteger, a smart wrapper around a sequence of u64 limbs, least-significant limb first.

Error

Field
The interface for a generic field.

FpParameters
A trait that defines parameters for a prime field.

Group

One
Defines a multiplicative identity element for Self.

PairingEngine

PrimeField
The interface for a prime field.

ProjectiveCurve
Projective representation of an elliptic curve point guaranteed to be in the correct prime order subgroup.

SquareRootField
The interface for a field that supports an efficient square-root operation.

UniformRand

Zero
Defines an additive identity element for Self.