Trait algebra::prelude::SquareRootField

pub trait SquareRootField: Field {
    fn legendre(&self) -> LegendreSymbol;

#[must_use]    fn sqrt(&self) -> Option<Self>;

    fn sqrt_in_place(&mut self) -> Option<&mut Self>;
}

The interface for a field that supports an efficient square-root operation.

Required methods

fn legendre(&self) -> LegendreSymbol

Returns the Legendre symbol.

#[must_use]fn sqrt(&self) -> Option<Self>

Returns the square root of self, if it exists.

fn sqrt_in_place(&mut self) -> Option<&mut Self>

Sets self to be the square root of self, if it exists.

Loading content...