algebra::ring::int_ring

Trait IntRing2k

pub trait IntRing2k:
    Display
    + Serialize
    + for<'a> Deserialize<'a>
    + WrappingAdd
    + WrappingSub
    + WrappingMul
    + WrappingNeg
    + Shl<usize, Output = Self>
    + Shr<usize, Output = Self>
    + Div<Output = Self>
    + Rem<Output = Self>
    + Not<Output = Self>
    + BitXor<Output = Self>
    + BitAnd<Output = Self>
    + BitOr<Output = Self>
    + BitXorAssign
    + BitAndAssign
    + BitOrAssign
    + From<bool>
    + TryInto<usize, Error: Debug>
    + Copy
    + Debug
    + Zero
    + One
    + Sized
    + Send
    + Sync
    + TryFrom<u128, Error: Debug>
    + TryFrom<u64, Error: Debug>
    + PartialEq
    + PartialOrd
    + 'static {
    const K: usize;
    const BYTES: usize;

    // Required methods
    fn from_reader<R: Read>(reader: R) -> Result<Self>;
    fn write<W: Write>(&self, writer: W) -> Result<()>;
    fn bits(&self) -> usize;
    fn cast_to_biguint(&self) -> BigUint;
    fn cast_from_biguint(biguint: &BigUint) -> Self;

    // Provided methods
    fn wrapping_add_assign(&mut self, rhs: &Self) { ... }
    fn wrapping_sub_assign(&mut self, rhs: &Self) { ... }
    fn wrapping_neg_inplace(&mut self) { ... }
    fn wrapping_mul_assign(&mut self, rhs: &Self) { ... }
}

Types implementing this trait can be used as elements of a ring Z_{2^k}

Required Associated Constants

const K: usize

Specifies the number of bits in this type

const BYTES: usize

Specifies the number of bytes used for storage in this type

Required Methods

fn from_reader<R: Read>(reader: R) -> Result<Self>

Reads a value of this type from a reader

fn write<W: Write>(&self, writer: W) -> Result<()>

Writes a value of this type to a writer

fn bits(&self) -> usize

Returns the effective number of bits (i.e., how many LSBs are set)

fn cast_to_biguint(&self) -> BigUint

Casts this type to a BigUint

fn cast_from_biguint(biguint: &BigUint) -> Self

Casts a BigUint to this type, removing any excess bits Thus if the value is larger than this type, it will be truncated

Provided Methods

fn wrapping_add_assign(&mut self, rhs: &Self)

a += b

fn wrapping_sub_assign(&mut self, rhs: &Self)

a -= b

fn wrapping_neg_inplace(&mut self)

a = -a

fn wrapping_mul_assign(&mut self, rhs: &Self)

a*= b

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.