Trait algebra::serialize::CanonicalDeserialize

pub trait CanonicalDeserialize {
    fn deserialize<R>(reader: R) -> Result<Self, SerializationError>
    where
        R: Read;

    fn deserialize_uncompressed<R>(
        reader: R
    ) -> Result<Self, SerializationError>
    where
        R: Read,
    { ... }

    fn deserialize_unchecked<R>(reader: R) -> Result<Self, SerializationError>
    where
        R: Read,
    { ... }
}

Deserializer in little endian format. This trait can be derived if all fields of a struct implement CanonicalDeserialize and the derive feature is enabled.

Example

// The `derive` feature must be set for the derivation to work.
use algebra_core::serialize::*;

#[derive(CanonicalDeserialize)]
struct TestStruct {
    a: u64,
    b: (u64, (u64, u64)),
}

If your code depends on algebra instead, the example works analogously when importing algebra::serialize::*.

Required methods

fn deserialize<R>(reader: R) -> Result<Self, SerializationError> where
R: Read,

Reads Self from reader.

Loading content...
Provided methods

fn deserialize_uncompressed<R>(reader: R) -> Result<Self, SerializationError> where
R: Read,

Reads Self from reader without compression.

fn deserialize_unchecked<R>(reader: R) -> Result<Self, SerializationError> where
R: Read,

Reads self from reader without compression, and without performing validity checks. Should be used only when the input is trusted.

Loading content...