Struct algebra::steenrod_evaluator::SteenrodEvaluator

pub struct SteenrodEvaluator {
    pub adem: AdemAlgebra,
    pub milnor: MilnorAlgebra,
}

Fields
adem: AdemAlgebramilnor: MilnorAlgebra
Implementations

impl SteenrodEvaluator

pub fn new(p: ValidPrime) -> Self

pub fn milnor_to_adem(
&self,
result: &mut FpVector,
coeff: u32,
degree: i32,
input: &FpVector
)

pub fn adem_to_milnor(
&self,
result: &mut FpVector,
coeff: u32,
degree: i32,
input: &FpVector
)

pub fn evaluate_algebra_adem(&self, input: &str) -> Result<(i32, FpVector)>

pub fn evaluate_algebra_milnor(&self, input: &str) -> Result<(i32, FpVector)>

pub fn evaluate_module_adem(
&self,
items: &str
) -> Result<BTreeMap<String, (i32, FpVector)>>

Returns

This returns a BTreeMap so that we get deterministic outputs for testing purposes

fn prime(&self) -> ValidPrime

fn compute_basis(&self, degree: i32)

fn dimension(&self, degree: i32) -> usize

fn evaluate_algebra_node(
&self,
output_degree: Option<i32>,
tree: AlgebraNode
) -> Result<(i32, FpVector)>

fn evaluate_basis_element(
&self,
output_degree: Option<i32>,
basis_elt: AlgebraBasisElt
) -> Result<(i32, FpVector)>

fn adem_to_milnor_on_basis(
&self,
result: &mut FpVector,
coeff: u32,
degree: i32,
idx: usize
)

Translate from the adem basis to the milnor basis, adding coeff times the result to result. This uses the fact that that $P^n = P(n)$ and $Q_1 = \beta$ and multiplies out the admissible monomial.

fn milnor_to_adem_on_basis(
&self,
result: &mut FpVector,
coeff: u32,
degree: i32,
idx: usize
)

fn milnor_to_adem_on_basis_2(
&self,
result: &mut FpVector,
coeff: u32,
degree: i32,
idx: usize
)

fn milnor_to_adem_on_basis_generic(
&self,
result: &mut FpVector,
coeff: u32,
degree: i32,
idx: usize
)

fn adem_q(&self, result: &mut FpVector, coeff: u32, qi: u32)

Express $Q_{qi}$ in the adem basis.

fn evaluate_p_or_b_list(&self, list: &[BocksteinOrSq]) -> (i32, FpVector)