Struct algebra::AdemAlgebra

pub struct AdemAlgebra {
    p: ValidPrime,
    generic: bool,
    unstable_enabled: bool,
    even_basis_table: OnceVec<Vec<AdemBasisElement>>,
    basis_table: OnceVec<Vec<AdemBasisElement>>,
    basis_element_to_index_map: OnceVec<FxHashMap<AdemBasisElement, usize>>,
    multiplication_table: OnceVec<Vec<Vec<FpVector>>>,
    excess_table: OnceVec<Vec<usize>>,
}

An Algebra implementing the Steenrod algebra, using the Adem basis.

Fields
p: ValidPrimegeneric: boolunstable_enabled: booleven_basis_table: OnceVec<Vec<AdemBasisElement>>basis_table: OnceVec<Vec<AdemBasisElement>>

degree -> index -> AdemBasisElement

basis_element_to_index_map: OnceVec<FxHashMap<AdemBasisElement, usize>>

degree -> AdemBasisElement -> index

multiplication_table: OnceVec<Vec<Vec<FpVector>>>

degree -> first square -> admissible sequence idx -> result

excess_table: OnceVec<Vec<usize>>
Implementations

impl AdemAlgebra

pub fn new(p: ValidPrime, unstable_enabled: bool) -> Self

Constructs a new AdemAlgebra.

pub fn generic(&self) -> bool

pub fn q(&self) -> i32

fn generate_basis_even(&self, max_degree: i32)

fn generate_basis2(&self, max_degree: i32)

fn generate_basis_generic(&self, max_degree: i32)

fn generate_basis_element_to_index_map(&self, max_degree: i32)

pub fn basis_element_from_index( &self, degree: i32, idx: usize, ) -> &AdemBasisElement

pub fn try_basis_element_to_index( &self, elt: &AdemBasisElement, ) -> Option<usize>

pub fn basis_element_to_index(&self, elt: &AdemBasisElement) -> usize

fn tail_of_basis_element_to_index( &self, elt: &mut AdemBasisElement, idx: u32, q: u32, ) -> usize

fn generate_multiplication_table_2(&self, max_degree: i32)

fn generate_multiplication_table_2_step( &self, table: &[Vec<FpVector>], n: i32, x: i32, idx: usize, ) -> FpVector

fn generate_multiplication_table_generic(&self, max_degree: i32)

fn generate_multiplication_table_generic_step( &self, table: &[Vec<FpVector>], n: i32, x: i32, idx: usize, ) -> FpVector

This function expresses $Sq^x$ (current) in terms of the admissible basis and returns the result as an FpVector, where (current) is the admissible monomial of degree $n - qx$ (so that $Sq^x)$ (current) has degree $n$) and index idx.

Here $Sq^x$ means $P^{x/2}$ if $x$ is even and $\beta P^{(x-1)/2}$ if $x$ is odd.

Note that x is always positive.

pub fn multiply_inner( &self, result: SliceMut<'_>, coeff: u32, r_degree: i32, r_index: usize, s_degree: i32, s_index: usize, excess: i32, )

pub fn make_mono_admissible( &self, result: SliceMut<'_>, coeff: u32, monomial: &mut AdemBasisElement, excess: i32, )

fn make_mono_admissible_2( &self, result: SliceMut<'_>, monomial: &mut AdemBasisElement, idx: i32, leading_degree: i32, excess: i32, stop_early: bool, )

Reduce a Steenrod monomial at the prime 2.

Arguments:

algebra - an Adem algebra. This would be a method of class AdemAlgebra.

result - Where we put the result

monomial - a not necessarily admissible Steenrod monomial which we will reduce. We destroy monomial->Ps.

idx - the only index to check for inadmissibility in the input (we assume that we’ve gotten our input as a product of two admissible sequences.)

leading_degree - the degree of the squares between 0 and idx (so of length idx + 1)

fn make_mono_admissible_generic( &self, result: SliceMut<'_>, coeff: u32, monomial: &mut AdemBasisElement, idx: i32, leading_degree: i32, excess: i32, stop_early: bool, )

fn decompose_basis_element_2( &self, degree: i32, idx: usize, ) -> Vec<(u32, (i32, usize), (i32, usize))>

fn decompose_basis_element_generic( &self, degree: i32, idx: usize, ) -> Vec<(u32, (i32, usize), (i32, usize))>

pub fn beps_pn(&self, e: u32, x: u32) -> (i32, usize)

impl AdemAlgebra

fn generate_excess_table(&self, max_degree: i32)