Trait algebra::Field

pub trait Field: 'static + ToBytes + FromBytes + CanonicalSerializeWithFlags + ConstantSerializedSize + CanonicalSerialize + CanonicalDeserializeWithFlags + CanonicalDeserialize + UniformRand + Clone + Eq + Ord + Default + Send + Copy + Sync + Add<Self, Output = Self, Output = Self> + for<'a> Add<&'a Self> + Sub<Self, Output = Self, Output = Self> + for<'a> Sub<&'a Self> + Mul<Self, Output = Self, Output = Self> + for<'a> Mul<&'a Self> + Div<Self, Output = Self, Output = Self> + for<'a> Div<&'a Self> + Neg<Output = Self> + AddAssign<Self> + for<'a> AddAssign<&'a Self> + SubAssign<Self> + for<'a> SubAssign<&'a Self> + MulAssign<Self> + for<'a> MulAssign<&'a Self> + DivAssign<Self> + for<'a> DivAssign<&'a Self> + Sum<Self> + for<'a> Sum<&'a Self> + Product<Self> + for<'a> Product<&'a Self> + Debug + Display + Hash + Zero + One {
    fn characteristic<'a>() -> &'a [u64];

#[must_use]    fn double(&self) -> Self;

    fn double_in_place(&mut self) -> &mut Self;

    fn from_random_bytes_with_flags(bytes: &[u8]) -> Option<(Self, u8)>;

#[must_use]    fn square(&self) -> Self;

    fn square_in_place(&mut self) -> &mut Self;

#[must_use]    fn inverse(&self) -> Option<Self>;

    fn inverse_in_place(&mut self) -> Option<&mut Self>;

    fn frobenius_map(&mut self, power: usize);

    fn from_random_bytes(bytes: &[u8]) -> Option<Self> { ... }

#[must_use]    fn pow<S>(&self, exp: S) -> Self
    where
        S: AsRef<[u64]>,
    { ... }
}

The interface for a generic field.

Required methods

fn characteristic<'a>() -> &'a [u64]

Returns the characteristic of the field.

#[must_use]fn double(&self) -> Self

Returns self + self.

fn double_in_place(&mut self) -> &mut Self

Doubles self in place.

fn from_random_bytes_with_flags(bytes: &[u8]) -> Option<(Self, u8)>

Returns a field element with an extra sign bit used for group parsing if the set of bytes forms a valid field element, otherwise returns None. This function is primarily intended for sampling random field elements from a hash-function or RNG output.

#[must_use]fn square(&self) -> Self

Returns self * self.

fn square_in_place(&mut self) -> &mut Self

Squares self in place.

#[must_use]fn inverse(&self) -> Option<Self>

Computes the multiplicative inverse of self if self is nonzero.

fn inverse_in_place(&mut self) -> Option<&mut Self>

fn frobenius_map(&mut self, power: usize)

Exponentiates this element by a power of the base prime modulus via the Frobenius automorphism.

Loading content...
Provided methods

fn from_random_bytes(bytes: &[u8]) -> Option<Self>

Returns a field element if the set of bytes forms a valid field element, otherwise returns None. This function is primarily intended for sampling random field elements from a hash-function or RNG output.

#[must_use]fn pow<S>(&self, exp: S) -> Self where
S: AsRef<[u64]>,

Exponentiates this element by a number represented with u64 limbs, least significant limb first.

Loading content...