Trait algebra::JsonAlgebra

pub trait JsonAlgebra: Algebra {
    fn prefix(&self) -> &str;

    fn json_to_basis(&self, json: &Value) -> Result<(i32, usize)>;

    fn json_from_basis(&self, degree: i32, idx: usize) -> Value;
}

Required methods

fn prefix(&self) -> &str

A name for the algebra to use in serialization operations.

fn json_to_basis(&self, json: &Value) -> Result<(i32, usize)>

Parses a basis element from JSON.

The representation is defined by the algebra itself, and is used by by module specifications to specify basis elements.

fn json_from_basis(&self, degree: i32, idx: usize) -> Value

Converts a basis element into JSON.