Trait algebra::ProjectiveCurve

pub trait ProjectiveCurve: 'static + ToBytes + FromBytes + UniformRand + Clone + Eq + From<Self::Affine> + Default + Send + Copy + Sync + Add<Self, Output = Self, Output = Self> + for<'a> Add<&'a Self> + Sub<Self, Output = Self, Output = Self> + for<'a> Sub<&'a Self> + Neg<Output = Self> + AddAssign<Self> + for<'a> AddAssign<&'a Self> + SubAssign<Self> + for<'a> SubAssign<&'a Self> + MulAssign<Self::ScalarField> + Sum<Self> + for<'a> Sum<&'a Self> + Debug + Display + Hash + Zero where
    <Self::Affine as AffineCurve>::Projective == Self,
    <Self::Affine as AffineCurve>::ScalarField == Self::ScalarField,
    <Self::Affine as AffineCurve>::BaseField == Self::BaseField, {
    type ScalarField: PrimeField + SquareRootField;
    type BaseField: Field;
    type Affine: AffineCurve + Into<Self> + From<Self>;

    const COFACTOR: &'static [u64];

#[must_use]    fn prime_subgroup_generator() -> Self;

    fn batch_normalization(v: &mut [Self]);

#[must_use]    fn is_normalized(&self) -> bool;

    fn double_in_place(&mut self) -> &mut Self;

    fn add_assign_mixed(&mut self, other: &Self::Affine);

    fn batch_normalization_into_affine(v: &[Self]) -> Vec<Self::Affine> { ... }

#[must_use]    fn double(&self) -> Self { ... }

    fn into_affine(&self) -> Self::Affine { ... }

    fn add_mixed(self, other: &Self::Affine) -> Self { ... }

    fn mul<S>(self, other: S) -> Self
    where
        S: Into<<Self::ScalarField as PrimeField>::BigInt>,
    { ... }
}

Projective representation of an elliptic curve point guaranteed to be in the correct prime order subgroup.

Associated Types

type ScalarField: PrimeField + SquareRootField

type BaseField: Field

type Affine: AffineCurve + Into<Self> + From<Self>

Loading content...
Associated Constants

const COFACTOR: &'static [u64]

Loading content...
Required methods

#[must_use]fn prime_subgroup_generator() -> Self

Returns a fixed generator of unknown exponent.

fn batch_normalization(v: &mut [Self])

Normalizes a slice of projective elements so that conversion to affine is cheap.

#[must_use]fn is_normalized(&self) -> bool

Checks if the point is already "normalized" so that cheap affine conversion is possible.

fn double_in_place(&mut self) -> &mut Self

Doubles this element in place.

fn add_assign_mixed(&mut self, other: &Self::Affine)

Set self to be self + other, where other: Self::Affine. This is usually faster than adding other in projective form.

Loading content...
Provided methods

fn batch_normalization_into_affine(v: &[Self]) -> Vec<Self::Affine>

Normalizes a slice of projective elements and outputs a vector containing the affine equivalents.

#[must_use]fn double(&self) -> Self

Doubles this element.

fn into_affine(&self) -> Self::Affine

Converts self into the affine representation.

fn add_mixed(self, other: &Self::Affine) -> Self

Set self to be self + other, where other: Self::Affine. This is usually faster than adding other in projective form.

fn mul<S>(self, other: S) -> Self where
S: Into<<Self::ScalarField as PrimeField>::BigInt>,

Performs scalar multiplication of this element.

Loading content...