Trait algebra::ToConstraintField

pub trait ToConstraintField<F> where
    F: Field, {
    fn to_field_elements(&self) -> Result<Vec<F>, Box<dyn Error + 'static>>;
}

Types that can be converted to a vector of F elements. Useful for specifying how public inputs to a constraint system should be represented inside that constraint system.

Required methods

fn to_field_elements(&self) -> Result<Vec<F>, Box<dyn Error + 'static>>

Loading content...