Crate algebra

Types and traits for working with various algebras and modules, with a focus on the Steenrod algebra and its modules.

Modules

adem_algebra

The Steenrod algebra using the Adem basis.

algebra 🔒

Traits describing algebras, and implementations thereof for different representations of the Steenrod algebra.

combinatorics

field

Finite fields over a prime.

milnor_algebra

module

pair_algebra

This implements the notion of a split pair algebra in the sense of https://arxiv.org/abs/2105.07628, whose notation we will use throughout. The Steenrod algebra admits a lift to a split pair algebra, which can be used to compute d2 differentials algorithmically.

steenrod_evaluator

steenrod_parser 🔒

This module includes code for parsing an expression in the Steenrod algebra into an abstract syntax tree.

Macros

dispatch_algebra

Structs

AdemAlgebra

An Algebra implementing the Steenrod algebra, using the Adem basis.

Field

$\mathbb{F}_p$, viewed as an Algebra over itself.

MilnorAlgebra

PolynomialAlgebraMonomial

PolynomialAlgebraTableEntry

Enums

AlgebraType

SteenrodAlgebra

SteenrodAlgebraBorrow

Traits

AdemAlgebraT

An algebra that can be viewed as an Adem algebra.

Algebra

A graded algebra over $\mathbb{F}_p$.

Bialgebra

An Algebra equipped with a coproduct operation that makes it into a bialgebra.

GeneratedAlgebra

An Algebra equipped with a distinguished presentation.

MilnorAlgebraT

MuAlgebra

An algebra that is maybe unstable. Every Algebra implements MuAlgebra by ignoring the excess parameter, and every UnstableAlgebra implements MuAlgebra. This makes it possible to write code that is generic over stable and unstable algebras.

PolynomialAlgebra

SteenrodAlgebraT

UnstableAlgebra

Functions

module_gens_from_json 🔒