algebra::algebra
Module adem_algebra

The Steenrod algebra using the Adem basis.

Structs

AdemAlgebra

An Algebra implementing the Steenrod algebra, using the Adem basis.

AdemBasisElement

An Adem basis element for the Steenrod algebra.

Enums

PorBockstein

A Steenrod power $P^i$, or a Bockstein $\beta^\varepsilon$.

Traits

AdemAlgebraT

An algebra that can be viewed as an Adem algebra.

Functions

shift_vec 🔒 ⚠

Shifts a Vec’s elements back by offset.