Trait algebra::algebra::algebra_trait::GeneratedAlgebra

pub trait GeneratedAlgebra: Algebra {
    fn generators(&self, degree: i32) -> Vec<usize>;

    fn string_to_generator<'a>(
        &self, 
        input: &'a str
    ) -> IResult<&'a str, (i32, usize)>;

    fn decompose_basis_element(
        &self, 
        degree: i32, 
        idx: usize
    ) -> Vec<(u32, (i32, usize), (i32, usize))>;

    fn generating_relations(
        &self, 
        degree: i32
    ) -> Vec<Vec<(u32, (i32, usize), (i32, usize))>>;

    fn generator_to_string(&self, degree: i32, idx: usize) -> String { ... }
}

An Algebra equipped with a distinguished presentation.

These data can be used to specify finite modules as the actions of the distinguished generators.

Required methods

fn generators(&self, degree: i32) -> Vec<usize>

Return generators in degree.

Generators are specified as basis element indices in that degree. The order of the list is not important.

This method need not be fast, because they will only be performed when constructing the module, and will often only involve low dimensional elements.

fn string_to_generator<'a>(
&self,
input: &'a str
) -> IResult<&'a str, (i32, usize)>

Parse input into a generator.

This function is a nom combinator.

This function MUST be inverse to string_to_generator (and not basis_element_to_string).

fn decompose_basis_element(
&self,
degree: i32,
idx: usize
) -> Vec<(u32, (i32, usize), (i32, usize))>

Decomposes an element into generators.

Given a basis element $A$, this function returns a list of triples $(c_i, A_i, B_i)$, where $A_i$ and $B_i$ are basis elements of strictly smaller degree than $A$, such that

$$ A = \sum_i c_i A_i B_i.$$

Combined with actions for generators, this allows us to recursively compute the action of an element on a module.

This method need not be fast, because they will only be performed when constructing the module, and will often only involve low dimensional elements.

fn generating_relations(
&self,
degree: i32
) -> Vec<Vec<(u32, (i32, usize), (i32, usize))>>

Returns relations that the algebra wants checked to ensure the consistency of module.

Relations are encoded as general multi-degree elements which are killed in the quotient: $$ \sum_i c_i \alpha_i \beta_i = 0. $$ where $c_i$ are coefficients and $\alpha_i$ and $\beta_i$ are basis elements of arbitrary degree.

Provided methods

fn generator_to_string(&self, degree: i32, idx: usize) -> String

Returns the name of a generator.

Note: idx is the index within degree’s basis, not the list returned by GeneratedAlgebra::generators().

By default, this function will forward to Algebra::basis_element_to_string(), but may be overridden if more concise names are available.

This function MUST be inverse to GeneratedAlgebra::string_to_generator().