Trait algebra::algebra::algebra_trait::MuAlgebra

pub trait MuAlgebra<const U: bool>: Algebra {
    fn dimension_unstable(&self, degree: i32, excess: i32) -> usize;
    fn multiply_basis_elements_unstable(
        &self,
        result: SliceMut<'_>,
        coeff: u32,
        r_degree: i32,
        r_index: usize,
        s_degree: i32,
        s_index: usize,
        excess: i32
    );
    fn multiply_basis_element_by_element_unstable(
        &self,
        result: SliceMut<'_>,
        coeff: u32,
        r_degree: i32,
        r_idx: usize,
        s_degree: i32,
        s: Slice<'_>,
        excess: i32
    );
    fn multiply_element_by_basis_element_unstable(
        &self,
        result: SliceMut<'_>,
        coeff: u32,
        r_degree: i32,
        r: Slice<'_>,
        s_degree: i32,
        s_idx: usize,
        excess: i32
    );
    fn multiply_element_by_element_unstable(
        &self,
        result: SliceMut<'_>,
        coeff: u32,
        r_degree: i32,
        r: Slice<'_>,
        s_degree: i32,
        s: Slice<'_>,
        excess: i32
    );
}

An algebra that is maybe unstable. Every Algebra implements MuAlgebra by ignoring the excess parameter, and every UnstableAlgebra implements MuAlgebra. This makes it possible to write code that is generic over stable and unstable algebras.

Required Methods

fn dimension_unstable(&self, degree: i32, excess: i32) -> usize

fn multiply_basis_elements_unstable(
&self,
result: SliceMut<'_>,
coeff: u32,
r_degree: i32,
r_index: usize,
s_degree: i32,
s_index: usize,
excess: i32
)

fn multiply_basis_element_by_element_unstable(
&self,
result: SliceMut<'_>,
coeff: u32,
r_degree: i32,
r_idx: usize,
s_degree: i32,
s: Slice<'_>,
excess: i32
)

fn multiply_element_by_basis_element_unstable(
&self,
result: SliceMut<'_>,
coeff: u32,
r_degree: i32,
r: Slice<'_>,
s_degree: i32,
s_idx: usize,
excess: i32
)

Computes the product r * s of a general element r and a basis element s, and adds the result to result.

Neither result nor r must be aligned.

fn multiply_element_by_element_unstable(
&self,
result: SliceMut<'_>,
coeff: u32,
r_degree: i32,
r: Slice<'_>,
s_degree: i32,
s: Slice<'_>,
excess: i32
)

Computes the product r * s of two general elements, and adds the result to result.

Neither result, s, nor r must be aligned.