Struct algebra::algebra::field::Field

pub struct Field {
    prime: ValidPrime,
}

$\mathbb{F}_p$, viewed as an Algebra over itself.

As an Algebra, a field is one-dimensional, with basis element 1. It is also trivially a coalgebra via the trivial diagonal comultiplication, and thus a Bialgebra.

Fields
prime: ValidPrime
Implementations

impl Field

pub fn new(p: ValidPrime) -> Self

Returns a new Field over the given prime p.