Module algebra::algebra::pair_algebra

This implements the notion of a split pair algebra in the sense of https://arxiv.org/abs/2105.07628, whose notation we will use throughout. The Steenrod algebra admits a lift to a split pair algebra, which can be used to compute d2 differentials algorithmically.

To keep the pair algebra business contained, we put the implementation of the Milnor algebra as a pair algebra in this file instead of milnor_algebra.rs.

Macros

sub 🔒

unsub 🔒

Structs

MilnorPairElement

Constants

AY_CACHE 🔒

Traits

PairAlgebra

A lift of an algebra to a split pair algebra. See module introduction for more.

Functions

a_y_cached 🔒

Compute $A(Sq(R), Y_{k, l})$ where $a = Sq(R)$. This queries the cache and computes it using a_y_inner if not available.

a_y_inner 🔒

Actually computes $A(a, Y_{k, l})$ and returns the result.

Type Definitions

HashMap 🔒