Module algebra::module

Modules

bcp

block_structure

bounded_module

dickson2

finite_dimensional_module

finite_module

finitely_presented_module

free_module

free_unstable_module

hom_module

homomorphism

kfpn

module_trait

polynomial_algebra_module

quotient_module

rpn

sum_module

tensor_module

truncated_module

zero_module

Structs

BCp

Dickson2

FDModule

FPModule

FreeModule

A free module.

FreeUnstableModule

HomModule

KFpn

ModuleFailedRelationError

OperationGeneratorPair

QuotientModule

Given a module module, this is the quotient of module by a collection of basis elements.

RealProjectiveSpace

This is $\mathbb{RP}_{\mathrm{min}}^{\mathrm{max}}$. The cohomology is the subquotient of $\mathbb{F}_2[x^\pm]$ given by elements of degree between min and max (inclusive)

SumModule

TensorModule

TruncatedModule

A module M where we quotient out everything above degree truncation

Enums

FiniteModule

Traits

BoundedModule

Module

PolynomialAlgebraModule

SteenrodModule

ZeroModule