Module algebra::module

Re-exports

pub use steenrod_module::SteenrodModule;

Modules

block_structure

finite_dimensional_module 🔒

finitely_presented_module 🔒

free_module 🔒

hom_module 🔒

homomorphism

module_trait 🔒

quotient_module 🔒

rpn 🔒

steenrod_module

suspension_module 🔒

tensor_module 🔒

zero_module 🔒

Structs

FDModule

FPModule

GeneratorData

HomModule

Given a module N and a free module M, this is the module Hom(M, N) as a module over the ground field. This requires N to be bounded, and is graded opposite to the usual grading so that Hom(M, N) is bounded below.

ModuleFailedRelationError

MuFreeModule

A free module.

OperationGeneratorPair

QuotientModule

A quotient of a module truncated below a fix degree.

RealProjectiveSpace

This is $\mathbb{RP}_{\mathrm{min}}^{\mathrm{max}}$. The cohomology is the subquotient of $\mathbb{F}_2[x^\pm]$ given by elements of degree between min and max (inclusive)

SuspensionModule

TensorModule

Traits

Module

A bounded below module over an algebra. To accommodate for infinite modules (e.g. modules in a free resolution), every module is potentially only define up to a degree. The extent to which the module is defined is kept track by two functions:

ZeroModule

Type Definitions

FreeModule

UnstableFreeModule