Struct algebra::module::free_module::MuFreeModule

pub struct MuFreeModule<const U: bool, A: MuAlgebra<U>> {
    algebra: Arc<A>,
    name: String,
    min_degree: i32,
    gen_names: OnceBiVec<Vec<String>>,
    gen_deg_idx_to_internal_idx: OnceBiVec<usize>,
    num_gens: OnceBiVec<usize>,
    basis_element_to_opgen: OnceBiVec<OnceVec<OperationGeneratorPair>>,
    generator_to_index: OnceBiVec<OnceVec<usize>>,
}

A free module.

A free module is uniquely determined by its list of generators. The generators are listed in increasing degrees, and the index in this list is the internal index.

Fields
algebra: Arc<A>name: Stringmin_degree: i32gen_names: OnceBiVec<Vec<String>>gen_deg_idx_to_internal_idx: OnceBiVec<usize>

degree -> internal index of first generator in degree

num_gens: OnceBiVec<usize>basis_element_to_opgen: OnceBiVec<OnceVec<OperationGeneratorPair>>generator_to_index: OnceBiVec<OnceVec<usize>>

degree -> internal_gen_idx -> the offset of the generator in degree

Implementations

impl<const U: bool, A: MuAlgebra<U>> MuFreeModule<U, A>

pub fn new(algebra: Arc<A>, name: String, min_degree: i32) -> Self

impl<const U: bool, A: MuAlgebra<U>> MuFreeModule<U, A>

pub fn gen_names(&self) -> &OnceBiVec<Vec<String>>

pub fn number_of_gens_in_degree(&self, degree: i32) -> usize

pub fn add_generators(
&self,
degree: i32,
num_gens: usize,
names: Option<Vec<String>>
)

pub fn internal_generator_offset(
&self,
degree: i32,
internal_gen_idx: usize
) -> usize

Given a generator (gen_deg, gen_idx), find the first index in degree degree with elements from the generator.

pub fn iter_gens(&self, degree: i32) -> impl Iterator<Item = (i32, usize)> + '_

Iterate the degrees and indices of each generator up to degree degree.

pub fn iter_gen_offsets<const N: usize>(
&self,
degree: [i32; N]
) -> impl Iterator<Item = GeneratorData<N>> + '_

Iterate the degrees and offsets of each generator up to degree degree.

pub fn generator_offset(&self, degree: i32, gen_deg: i32, gen_idx: usize) -> usize

Given a generator (gen_deg, gen_idx), find the first index in degree degree with elements from the generator.

pub fn operation_generator_to_index(
&self,
op_deg: i32,
op_idx: usize,
gen_deg: i32,
gen_idx: usize
) -> usize

pub fn index_to_op_gen(
&self,
degree: i32,
index: usize
) -> &OperationGeneratorPair

pub fn extend_by_zero(&self, degree: i32)

pub fn element_to_string_pretty(&self, s: u32, t: i32, vec: Slice<'_>) -> String

A version of element_to_string that names the generator as x_(t - s, s, idx). The input s only affects how the output is displayed.

pub fn slice_vector<'a>(
&self,
degree: i32,
gen_degree: i32,
gen_index: usize,
v: Slice<'a>
) -> Slice<'a>

Given a vector that represents an element in degree degree, slice it to the part that represents the terms that correspond to the specified generator.

pub fn iter_slices<'a>(
&'a self,
degree: i32,
slice: Slice<'a>
) -> impl Iterator<Item = (i32, usize, i32, Slice<'a>)> + 'a

Given an element in a degree, iterate through the slices corresponding to each generator. Each item of the iterator is (gen_degree, gen_index, op_degree, slice). This skips slices that are zero length.