Module algebra::module::homomorphism

Modules

free_module_homomorphism 🔒

full_module_homomorphism 🔒

generic_zero_homomorphism 🔒

hom_pullback 🔒

quotient_homomorphism 🔒

Structs

FullModuleHomomorphism

A ModuleHomomorphism that simply records the matrix of the homomorphism in every degree. This is currently rather bare bones.

GenericZeroHomomorphism

HomPullback

Given a map $\mathtt{map}: A \to B$ and hom modules $\mathtt{source} = \Hom(B, X)$, $\mathtt{target} = \Hom(A, X)$, produce the induced pullback map $\Hom(B, X) \to \Hom(A, X)$.

MuFreeModuleHomomorphism

QuotientHomomorphism

QuotientHomomorphismSource

Traits

IdentityHomomorphism

ModuleHomomorphism

Each ModuleHomomorphism may come with auxiliary data, namely the kernel, image and quasi_inverse at each degree (the quasi-inverse is a map that is a right inverse when restricted to the image). These are computed via ModuleHomomorphism::compute_auxiliary_data_through_degree and retrieved through ModuleHomomorphism::kernel, ModuleHomomorphism::quasi_inverse and ModuleHomomorphism::image.

ZeroHomomorphism

Type Aliases

FreeModuleHomomorphism

UnstableFreeModuleHomomorphism