Struct algebra::module::homomorphism::full_module_homomorphism::FullModuleHomomorphism

pub struct FullModuleHomomorphism<S: Module, T: Module<Algebra = S::Algebra> = S> {
    source: Arc<S>,
    target: Arc<T>,
    degree_shift: i32,
    matrices: OnceBiVec<Matrix>,
    quasi_inverses: OnceBiVec<QuasiInverse>,
    kernels: OnceBiVec<Subspace>,
    images: OnceBiVec<Subspace>,
}

A ModuleHomomorphism that simply records the matrix of the homomorphism in every degree. This is currently rather bare bones.

Fields
source: Arc<S>target: Arc<T>degree_shift: i32matrices: OnceBiVec<Matrix>

The matrices of the module homomorphism. Unspecified matrices are assumed to be zero

quasi_inverses: OnceBiVec<QuasiInverse>kernels: OnceBiVec<Subspace>images: OnceBiVec<Subspace>
Implementations

impl<A, S, T> FullModuleHomomorphism<S, T>
where A: Algebra, S: Module<Algebra = A>, T: Module<Algebra = A>,

pub fn new(source: Arc<S>, target: Arc<T>, degree_shift: i32) -> Self

pub fn from_matrices( source: Arc<S>, target: Arc<T>, degree_shift: i32, matrices: BiVec<Matrix> ) -> Self

pub fn from<F: ModuleHomomorphism<Source = S, Target = T>>(f: &F) -> Self

pub fn replace_source<S_: Module<Algebra = A>>( self, source: Arc<S_> ) -> FullModuleHomomorphism<S_, T>

This function replaces the source of the ModuleHomomorphism and does nothing else. This is useful for changing the type of the source (but not the mathematical module itself).

pub fn replace_target<T_: Module<Algebra = A>>( self, target: Arc<T_> ) -> FullModuleHomomorphism<S, T_>

See replace_source