Trait algebra::module::module_trait::Module

pub trait Module: Send + Sync + 'static {
    type Algebra: Algebra;
    fn algebra(&self) -> Arc<Self::Algebra>;

    fn name(&self) -> String;

    fn min_degree(&self) -> i32;

    fn dimension(&self, degree: i32) -> usize;

    fn act_on_basis(
        &self, 
        result: &mut FpVector, 
        coeff: u32, 
        op_degree: i32, 
        op_index: usize, 
        mod_degree: i32, 
        mod_index: usize
    );

    fn basis_element_to_string(&self, degree: i32, idx: usize) -> String;

    fn compute_basis(&self, _degree: i32) { ... }

    fn is_unit(&self) -> bool { ... }

    fn prime(&self) -> ValidPrime { ... }

    fn borrow_output(&self) -> bool { ... }

    fn act_on_basis_borrow(
        &self, 
        _op_degree: i32, 
        _op_index: usize, 
        _mod_degree: i32, 
        _mod_index: usize
    ) -> &FpVector { ... }

    fn act(
        &self, 
        result: &mut FpVector, 
        coeff: u32, 
        op_degree: i32, 
        op_index: usize, 
        input_degree: i32, 
        input: &FpVector
    ) { ... }

    fn act_by_element(
        &self, 
        result: &mut FpVector, 
        coeff: u32, 
        op_degree: i32, 
        op: &FpVector, 
        input_degree: i32, 
        input: &FpVector
    ) { ... }

    fn generator_list_string(&self, degree: i32) -> String { ... }

    fn element_to_string(&self, degree: i32, element: &FpVector) -> String { ... }

    fn truncate_to_fd_module(
        self: Arc<Self>, 
        max_deg: i32
    ) -> FDModule<Self::Algebra> { ... }
}

Associated Types

type Algebra: Algebra

Loading content...
Required methods

fn algebra(&self) -> Arc<Self::Algebra>

fn name(&self) -> String

fn min_degree(&self) -> i32

fn dimension(&self, degree: i32) -> usize

fn act_on_basis(
&self,
result: &mut FpVector,
coeff: u32,
op_degree: i32,
op_index: usize,
mod_degree: i32,
mod_index: usize
)

fn basis_element_to_string(&self, degree: i32, idx: usize) -> String

Loading content...
Provided methods

fn compute_basis(&self, _degree: i32)

fn is_unit(&self) -> bool

Whether this is the unit module.

fn prime(&self) -> ValidPrime

fn borrow_output(&self) -> bool

Whether act_on_basis_borrow is available.

fn act_on_basis_borrow(
&self,
_op_degree: i32,
_op_index: usize,
_mod_degree: i32,
_mod_index: usize
) -> &FpVector

Returns a borrow of the value of the corresponding action on the basis element. This FpVector must be "pure", i.e. it is not sliced and the limbs are zero in indices greater than the dimension of the vector.

fn act(
&self,
result: &mut FpVector,
coeff: u32,
op_degree: i32,
op_index: usize,
input_degree: i32,
input: &FpVector
)

fn act_by_element(
&self,
result: &mut FpVector,
coeff: u32,
op_degree: i32,
op: &FpVector,
input_degree: i32,
input: &FpVector
)

fn generator_list_string(&self, degree: i32) -> String

fn element_to_string(&self, degree: i32, element: &FpVector) -> String

fn truncate_to_fd_module(
self: Arc<Self>,
max_deg: i32
) -> FDModule<Self::Algebra>

This truncates the module to max_dim and represents it as an FDModule. This retains the original name of the module

Loading content...