Trait algebra::module::module_trait::Module

pub trait Module: Display + Send + Sync {
    type Algebra: Algebra;

    fn algebra(&self) -> Arc<Self::Algebra>;

    fn min_degree(&self) -> i32;

    fn max_computed_degree(&self) -> i32;

    fn dimension(&self, degree: i32) -> usize;

    fn act_on_basis(
        &self, 
        result: SliceMut<'_>, 
        coeff: u32, 
        op_degree: i32, 
        op_index: usize, 
        mod_degree: i32, 
        mod_index: usize
    );

    fn basis_element_to_string(&self, degree: i32, idx: usize) -> String;

    fn compute_basis(&self, _degree: i32) { ... }

    fn is_unit(&self) -> bool { ... }

    fn prime(&self) -> ValidPrime { ... }

    fn borrow_output(&self) -> bool { ... }

    fn act_on_basis_borrow(
        &self, 
        _op_degree: i32, 
        _op_index: usize, 
        _mod_degree: i32, 
        _mod_index: usize
    ) -> &FpVector { ... }

    fn act(
        &self, 
        result: SliceMut<'_>, 
        coeff: u32, 
        op_degree: i32, 
        op_index: usize, 
        input_degree: i32, 
        input: Slice<'_>
    ) { ... }

    fn act_by_element(
        &self, 
        result: SliceMut<'_>, 
        coeff: u32, 
        op_degree: i32, 
        op: Slice<'_>, 
        input_degree: i32, 
        input: Slice<'_>
    ) { ... }

    fn act_by_element_on_basis(
        &self, 
        result: SliceMut<'_>, 
        coeff: u32, 
        op_degree: i32, 
        op: Slice<'_>, 
        input_degree: i32, 
        input_index: usize
    ) { ... }

    fn basis_string_list(&self, degree: i32) -> Vec<String> { ... }

    fn element_to_string(&self, degree: i32, element: Slice<'_>) -> String { ... }

    fn truncate_to_fd_module(
        self: Arc<Self>, 
        max_deg: i32
    ) -> FDModule<Self::Algebra> { ... }

    fn check_relation(
        &self, 
        result: &mut FpVector, 
        scratch: &mut FpVector, 
        outer_op_degree: i32, 
        outer_op_index: usize, 
        inner_op_degree: i32, 
        inner_op_index: usize, 
        module_degree: i32, 
        module_index: usize
    ) { ... }

    fn check_relations(
        &self, 
        max_degree: i32
    ) -> Vec<((i32, usize, i32, usize, i32, usize), FpVector)> { ... }

    fn test_relations(&self, max_degree: i32, max_failures_to_display: usize) { ... }

}

Associated Types

type Algebra: Algebra

Required methods

fn algebra(&self) -> Arc<Self::Algebra>

fn min_degree(&self) -> i32

fn max_computed_degree(&self) -> i32

The maximum t for which the module is defined at t.

fn dimension(&self, degree: i32) -> usize

fn act_on_basis(
&self,
result: SliceMut<'_>,
coeff: u32,
op_degree: i32,
op_index: usize,
mod_degree: i32,
mod_index: usize
)

fn basis_element_to_string(&self, degree: i32, idx: usize) -> String

Provided methods

fn compute_basis(&self, _degree: i32)

fn is_unit(&self) -> bool

Whether this is the unit module.

fn prime(&self) -> ValidPrime

fn borrow_output(&self) -> bool

Whether act_on_basis_borrow is available.

fn act_on_basis_borrow(
&self,
_op_degree: i32,
_op_index: usize,
_mod_degree: i32,
_mod_index: usize
) -> &FpVector

Returns a borrow of the value of the corresponding action on the basis element.

fn act(
&self,
result: SliceMut<'_>,
coeff: u32,
op_degree: i32,
op_index: usize,
input_degree: i32,
input: Slice<'_>
)

The length of input need not be equal to the dimension of the module in said degree. Missing entries are interpreted to be 0, while extra entries must be zero.

This flexibility is useful when resolving to a stem. The point is that we have elements in degree t that are guaranteed to not contain generators of degree t, and we don’t know what generators will be added in degree t yet.

fn act_by_element(
&self,
result: SliceMut<'_>,
coeff: u32,
op_degree: i32,
op: Slice<'_>,
input_degree: i32,
input: Slice<'_>
)

fn act_by_element_on_basis(
&self,
result: SliceMut<'_>,
coeff: u32,
op_degree: i32,
op: Slice<'_>,
input_degree: i32,
input_index: usize
)

fn basis_string_list(&self, degree: i32) -> Vec<String>

fn element_to_string(&self, degree: i32, element: Slice<'_>) -> String

fn truncate_to_fd_module(
self: Arc<Self>,
max_deg: i32
) -> FDModule<Self::Algebra>

This truncates the module to max_dim and represents it as an FDModule. This retains the original name of the module

fn check_relation(
&self,
result: &mut FpVector,
scratch: &mut FpVector,
outer_op_degree: i32,
outer_op_index: usize,
inner_op_degree: i32,
inner_op_index: usize,
module_degree: i32,
module_index: usize
)

op1(op2(x)) - (op1*op2)(x)

fn check_relations(
&self,
max_degree: i32
) -> Vec<((i32, usize, i32, usize, i32, usize), FpVector)>

Input: degree through which to check. Output: Vec of discrepancies.

fn test_relations(&self, max_degree: i32, max_failures_to_display: usize)