Struct algebra::module::rpn::RealProjectiveSpace

pub struct RealProjectiveSpace<A: Algebra> {
    algebra: Arc<A>,
    pub min: i32,
    pub max: Option<i32>,
    pub clear_bottom: bool,
}

This is $\mathbb{RP}_{\mathrm{min}}^{\mathrm{max}}$. The cohomology is the subquotient of $\mathbb{F}_2[x^\pm]$ given by elements of degree between min and max (inclusive)

The clear_bottom option, if selected, mods out by the elements in the $A(2)$ submodule generated by degrees less than min. This is useful for approximating $\mathrm{tmf} \wedge \mathbb{RP}_{-\infty}^n$, c.f. Proposition 2.2 of Bailey and Ricka. Note that this quotient always has minimum degree -1 mod 8.

Fields
algebra: Arc<A>min: i32max: Option<i32>clear_bottom: bool
Implementations

impl<A: Algebra> RealProjectiveSpace<A>
where for<'a> &'a A: TryInto<&'a SteenrodAlgebra>,

pub fn new( algebra: Arc<A>, min: i32, max: Option<i32>, clear_bottom: bool ) -> Self

impl<A: Algebra> RealProjectiveSpace<A>

pub fn from_json(algebra: Arc<A>, json: &Value) -> Result<Self>

pub fn to_json(&self, json: &mut Value)