Struct algebra::module::FDModule

pub struct FDModule<A: Algebra> {
    algebra: Arc<A>,
    pub name: String,
    graded_dimension: BiVec<usize>,
    gen_names: BiVec<Vec<String>>,
    actions: BiVec<BiVec<Vec<Vec<FpVector>>>>,
}

Fields
algebra: Arc<A>name: Stringgraded_dimension: BiVec<usize>gen_names: BiVec<Vec<String>>actions: BiVec<BiVec<Vec<Vec<FpVector>>>>
Implementations

impl<A: Algebra> FiniteDimensionalModule<A>

pub fn test_equal(&self, other: &Self) -> Result<(), String>

impl<A: Algebra> FiniteDimensionalModule<A>

pub fn new(
algebra: Arc<A>,
name: String,
graded_dimension: BiVec<usize>
) -> Self

pub fn set_basis_element_name(&mut self, degree: i32, idx: usize, name: String)

fn allocate_actions(
algebra: &Arc<A>,
graded_dimension: &BiVec<usize>
) -> BiVec<BiVec<Vec<Vec<FpVector>>>>

pub fn add_generator(&mut self, degree: i32, name: String)

pub fn string_to_basis_element(&self, string: &str) -> Option<(i32, usize)>

pub fn set_action_vector(
&mut self,
operation_degree: i32,
operation_idx: usize,
input_degree: i32,
input_idx: usize,
output: &FpVector
)

pub fn set_action(
&mut self,
operation_degree: i32,
operation_idx: usize,
input_degree: i32,
input_idx: usize,
output: Vec<u32>
)

pub fn action(
&self,
operation_degree: i32,
operation_idx: usize,
input_degree: i32,
input_idx: usize
) -> &FpVector

This function will panic if you call it with input such that module.dimension(input_degree + operation_degree) = 0.

pub fn action_mut(
&mut self,
operation_degree: i32,
operation_idx: usize,
input_degree: i32,
input_idx: usize
) -> &mut FpVector

This function will panic if you call it with input such that module.dimension(input_degree + operation_degree) = 0.

impl<A: JsonAlgebra + GeneratedAlgebra> FiniteDimensionalModule<A>

fn module_gens_from_json(
gens: &Value
) -> (BiVec<usize>, BiVec<Vec<String>>, HashMap<String, (i32, usize)>)

pub fn from_json(algebra: Arc<A>, json: &Value) -> Result<Self>

pub fn to_json(&self, json: &mut Value)

pub fn parse_action(
&mut self,
gen_to_idx: &HashMap<String, (i32, usize)>,
entry_: &str,
overwrite: bool
) -> Result<()>

pub fn parse_element(
&self,
entry: &str,
degree: i32,
result: SliceMut<'_>
) -> Result<()>

fn take_element(i: &str) -> IResult<&str, (u32, &str)>

pub fn check_validity(
&self,
input_deg: i32,
output_deg: i32
) -> Result<(), ModuleFailedRelationError>

pub fn extend_actions(&mut self, input_deg: i32, output_deg: i32)

pub fn minimal_actions_to_json(&self) -> Value

pub fn to_minimal_json(&self) -> Value

pub fn actions_to_json(&self) -> Value

pub fn gens_to_json(&self) -> Value