Module algebra::steenrod_parser

This module includes code for parsing an expression in the Steenrod algebra into an abstract syntax tree.

Enums

AlgebraBasisElt

AlgebraNode

BocksteinOrSq

Functions

algebra_expr 🔒

algebra_factor 🔒

algebra_generator 🔒

algebra_term 🔒

brackets 🔒

Surround with brackets

convert_error 🔒

digits 🔒

fold_separated 🔒

module_expr 🔒

module_generator 🔒

module_term 🔒

p_or_sq 🔒

parse_algebra

parse_module

scalar 🔒

space 🔒

Pad both ends with whitespace

Type Definitions

IResult 🔒

ModuleNode