Struct algebra::MilnorAlgebra

pub struct MilnorAlgebra {
    profile: MilnorProfile,
    p: ValidPrime,
    generic: bool,
    unstable_enabled: bool,
    ppart_table: OnceVec<Vec<PPart>>,
    basis_table: OnceVec<Vec<MilnorBasisElement>>,
    excess_table: OnceVec<Vec<usize>>,
    basis_element_to_index_map: OnceVec<HashMap<MilnorBasisElement, usize>>,
    multiplication_table: OnceVec<OnceVec<Vec<Vec<FpVector>>>>,
}

Fields
profile: MilnorProfilep: ValidPrimegeneric: boolunstable_enabled: boolppart_table: OnceVec<Vec<PPart>>

This is a list of possible P(R) of each degree, where ppart_table[i] contains elements of degree q * i.

basis_table: OnceVec<Vec<MilnorBasisElement>>

A list of all basis elements of each degree, constructed from Self::ppart_table

excess_table: OnceVec<Vec<usize>>basis_element_to_index_map: OnceVec<HashMap<MilnorBasisElement, usize>>

degree -> MilnorBasisElement -> index

multiplication_table: OnceVec<OnceVec<Vec<Vec<FpVector>>>>

source_deg -> target_deg -> source_op -> target_op

Implementations

impl MilnorAlgebra

pub fn new(p: ValidPrime, unstable_enabled: bool) -> Self

pub fn new_with_profile(
p: ValidPrime,
profile: MilnorProfile,
unstable_enabled: bool
) -> Self

pub fn generic(&self) -> bool

pub fn q(&self) -> i32

pub fn profile(&self) -> &MilnorProfile

pub fn basis_element_from_index(
&self,
degree: i32,
idx: usize
) -> &MilnorBasisElement

pub fn try_basis_element_to_index(
&self,
elt: &MilnorBasisElement
) -> Option<usize>

pub fn basis_element_to_index(&self, elt: &MilnorBasisElement) -> usize

pub fn ppart_table(&self, t: i32) -> &[PPart]

Gives a list of PPart’s in degree t.

impl MilnorAlgebra

fn compute_ppart(&self, max_degree: i32)

fn generate_basis_generic(&self, max_degree: i32)

fn generate_basis_2(&self, max_degree: i32)

fn generate_excess_table(&self, max_degree: i32)

impl MilnorAlgebra

fn try_beps_pn(&self, e: u32, x: PPartEntry) -> Option<(i32, usize)>

pub fn beps_pn(&self, e: u32, x: PPartEntry) -> (i32, usize)

Return the degree and index of $Q_1^e P(x)$.

fn multiply_qpart(
&self,
m1: &MilnorBasisElement,
f: u32
) -> Vec<(u32, MilnorBasisElement)>

pub fn multiply(
&self,
res: SliceMut<'_>,
coef: u32,
m1: &MilnorBasisElement,
m2: &MilnorBasisElement
)

pub fn multiply_with_allocation(
&self,
res: SliceMut<'_>,
coef: u32,
m1: &MilnorBasisElement,
m2: &MilnorBasisElement,
excess: i32,
allocation: PPartAllocation
) -> PPartAllocation

pub fn multiply_basis_by_element(
&self,
res: SliceMut<'_>,
coef: u32,
m1: &MilnorBasisElement,
s_deg: i32,
s: Slice<'_>
)

fn multiply_basis_by_element_with_allocation(
&self,
res: SliceMut<'_>,
coef: u32,
m1: &MilnorBasisElement,
s_deg: i32,
s: Slice<'_>,
allocation: PPartAllocation
) -> PPartAllocation

impl MilnorAlgebra

fn decompose_basis_element_qpart(
&self,
degree: i32,
idx: usize
) -> Vec<(u32, (i32, usize), (i32, usize))>

fn decompose_basis_element_ppart(
&self,
degree: i32,
idx: usize
) -> Vec<(u32, (i32, usize), (i32, usize))>

impl MilnorAlgebra

fn increment_p_part(element: &mut PPart, max: &[PPartEntry]) -> bool

Returns true if the new element is not within the bounds